cmd/geth: remove subcmd hbss2pbss and insecure-prune-all (#3346)
