+++ core/rawdb/database.go
@@ -587,28 +587,6 @@ func AncientInspect(db ethdb.Database) error {
 	return nil
 }
 
-func PruneHashTrieNodeInDataBase(db ethdb.Database) error {
-	it := db.NewIterator([]byte{}, []byte{})
-	defer it.Release()
-
-	total_num := 0
-	for it.Next() {
-		var key = it.Key()
-		switch {
-		case IsLegacyTrieNode(key, it.Value()):
-			db.Delete(key)
-			total_num++
-			if total_num%100000 == 0 {
-				log.Info("Pruning hash-base state trie nodes", "Complete progress: ", total_num)
-			}
-		default:
-			continue
-		}
-	}
-	log.Info("Pruning hash-base state trie nodes", "Complete progress", total_num)
-	return nil
-}
-
 type DataType int
 
 const (
