+++ trie/trie.go
@@ -711,3 +711,14 @@ func (t *Trie) Reset() {
 	t.tracer.reset()
 	t.committed = false
 }
+
+func (t *Trie) resloveWithoutTrack(n node, prefix []byte) (node, error) {
+	if n, ok := n.(hashNode); ok {
+		blob, err := t.reader.node(prefix, common.BytesToHash(n))
+		if err != nil {
+			return nil, err
+		}
+		return mustDecodeNode(n, blob), nil
+	}
+	return n, nil
+}
